Changing so no city with the same name can be added to Database.

+++ src/main/java/org/easytravel/service/implementation/CityServiceImplementation.java
@@ -3,6 +3,7 @@ package org.easytravel.service.implementation;
 import lombok.AllArgsConstructor;
 import org.easytravel.model.City;
 import org.easytravel.model.exceptions.CityNotFoundException;
+import org.easytravel.model.exceptions.CityWithThatNameAlreadyExists;
 import org.easytravel.repository.CityRepository;
 import org.easytravel.service.CityService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,13 +15,16 @@ import java.util.List;
 @AllArgsConstructor
 public class CityServiceImplementation implements CityService {
 
-   @Autowired
-   private final CityRepository cityRepository;
+    @Autowired
+    private final CityRepository cityRepository;
 
     @Override
     public City createCity(String cityName) {
+        if (cityRepository.existsByName(cityName)) {
+            throw new CityWithThatNameAlreadyExists(cityName);
+        }
         return cityRepository.save(
-            new City(cityName = cityName)
+                new City(cityName = cityName)
         );
     }
 
@@ -43,12 +47,16 @@ public class CityServiceImplementation implements CityService {
 
     @Override
     public void deleteCity(City city) {
-        cityRepository.delete(city);
+        if (cityRepository.existsById(city.getId())) {
+            cityRepository.delete(city);
+        }
     }
 
     @Override
     public void deleteCity(Long cityId) {
-        cityRepository.delete(cityRepository.findById(cityId).orElseThrow(CityNotFoundException::new));
+        if (cityRepository.existsById(cityId)) {
+            cityRepository.delete(cityRepository.findById(cityId).orElseThrow(CityNotFoundException::new));
+        }
     }
 
     @Override
